refactor: organize schema package
