+++ app/db/session.py
@@ -5,7 +5,7 @@ from sqlalchemy.orm import sessionmaker
 
 from app.core.settings import settings
 
-SQLALCHEMY_DATABASE_URL = settings.DB_URL
+SQLALCHEMY_DATABASE_URL = settings.db_url
 engine = create_engine(SQLALCHEMY_DATABASE_URL)
 
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
