+++ app/schema/__init__.py
@@ -1,0 +1,3 @@
+import sqlalchemy
+
+metadata = sqlalchemy.MetaData()
