+++ app/schema/statistic.py
@@ -3,6 +3,7 @@ from enum import Enum
 from typing import Optional
 
 from sqlalchemy import Column, DateTime
+from sqlalchemy import Enum as SAEnum
 from sqlalchemy.dialects.postgresql import JSON
 from sqlalchemy.sql import func
 from sqlmodel import Field, SQLModel
@@ -15,6 +16,8 @@ class StatisticType(Enum):
 
 class Statistic(SQLModel, table=True):
     id: Optional[int] = Field(default=None, primary_key=True)
-    type: StatisticType = Field(default=StatisticType.MSME_OPT_IN_STATISTICS)
+    type: StatisticType = Field(sa_column=Column(SAEnum(StatisticType, name="user_type")))
     data: dict = Field(default={}, sa_column=Column(JSON))
-    updated_at: datetime = Field(sa_column=Column(DateTime(timezone=True), nullable=False, onupdate=func.now()))
+    updated_at: Optional[datetime] = Field(
+        sa_column=Column(DateTime(timezone=True), nullable=False, onupdate=func.now())
+    )
